+++ Meraki.Api/Data/OrganizationPrivacy.cs
@@ -1,0 +1,15 @@
+namespace Meraki.Api.Data;
+
+/// <summary>
+/// Organization privacy settings
+/// </summary>
+[ApiAccessReadOnlyClass]
+[DataContract]
+public class OrganizationPrivacy
+{
+	/// <summary>
+	/// Default privacy settings
+	/// </summary>
+	[DataMember(Name = "byDefault")]
+	public OrganizationPrivacyByDefault ByDefault { get; set; } = new();
+}